requirements.txt
amqp==2.5.2  # BSD-license
boto3==1.10.*  # Apache 2.0-license
celery==4.4.0  # BSD-license
coreapi==2.3.3  # BSD-license
coreschema==0.0.4  # BSD-license
cryptography==2.8  # BSD-license
Django==2.2.9  # BSD-license
django-cas-ng==3.6.0  # BSD-license
django-celery-beat==1.5.0  # BSD-license
django-celery-results==1.1.2  # BSD-license
django-cors-headers==2.5.3  # MIT-license
django-crispy-forms==1.8.1  # MIT-license
django-csp==3.5  # BSD-license
django-filter==2.2.0  # BSD-license
django-guardian==1.5.1  # BSD-license
django-pylibmc==0.6.1  # BSD-license
django-ratelimit==2.0.0  # Apache 2.0-license
django-rest-swagger==2.2.0  # FreeBSD-license
django-simple-history==2.7.3  # BSD-license
djangorestframework==3.11.0  # BSD-license
djangorestframework-guardian==0.3.0  # BSD-license
drf-nested-routers==0.91  # Apache 2.0-license
flake8==3.7.9  # MIT-license
flower==0.9.3  # BSD-license
gevent==1.4.0  # MIT-license
gunicorn==20.0.4  # MIT-license
itypes==1.1.0  # BSD-license
Jinja2==2.10.3  # BSD-license
kombu==4.6.7  # BSD-license
MarkupSafe==1.1.1  # BSD-license
mccabe==0.6.1  # Expat license (same as MIT)
openapi-codec==1.3.2  # BSD-license
psycopg2-binary==2.8.4  # LGPL with exceptions or ZPL
pycodestyle==2.5.0  # Expat license (same as MIT)
pyflakes==2.1.1  # MIT-license
pylibmc==1.6.1  # BSD 3-Clause License
PyYAML==5.2  # MIT-license
requests==2.22.0  # Apache 2.0-license
simplejson==3.17.0  # MIT-license
uritemplate==3.0.0  # BSD 3-Clause License or Apache License, Version 2.0
responses==0.10.8  # Mocks. Apache 2.0-license
